Show overwrite dialog

+++ recce/server.py
@@ -183,6 +183,11 @@ async def get_info():
         state = context.export_state()
 
     support_tasks = context.support_tasks()
+    if context.state_loader and context.state_loader.state_file:
+        filename = os.path.basename(context.state_loader.state_file)
+    else:
+        filename = None
+
     try:
         info = {
             'adapter_type': context.adapter_type,
@@ -196,7 +201,7 @@ async def get_info():
             'demo': bool(demo),
             'cloud_mode': context.state_loader.cloud_mode,
             'file_mode': context.state_loader.state_file is not None,
-            'filename': os.path.basename(context.state_loader.state_file),
+            'filename': filename,
             'support_tasks': support_tasks,
         }
 
@@ -275,70 +280,62 @@ async def save_handler():
         raise HTTPException(status_code=400, detail=e.message)
 
 
-class SaveAsInput(BaseModel):
+class SaveAsOrRenameInput(BaseModel):
+    # The filename. The filename should not contain directory.
     filename: Optional[str] = None
+    # Overwrite the file if it exists
+    overwrite: Optional[bool] = False
+
+
+def saveas_or_rename(input: SaveAsOrRenameInput, rename: bool = False):
+    context = default_context()
+    state_loader = context.state_loader
+    if state_loader.cloud_mode:
+        raise RecceException('Cloud mode does not support rename')
+
+    new_filename = input.filename
+    if os.path.dirname(new_filename):
+        raise RecceException('The new filename should not contain directory')
+    if not new_filename.endswith('.json'):
+        raise RecceException('The new filename should end with .json')
+
+    old_path = state_loader.state_file
+    if old_path:
+        old_dir = os.path.dirname(state_loader.state_file)
+        old_filename = os.path.basename(state_loader.state_file)
+        if old_filename == new_filename:
+            raise RecceException('The new filename is the same as the current filename')
+        new_path = os.path.join(old_dir, new_filename)
+    else:
+        new_path = new_filename
+
+    if os.path.exists(new_path) and not input.overwrite:
+        raise HTTPException(status_code=409, detail=f'The file {new_filename} already exists')
+
+    state_loader.state_file = new_path
+    context.sync_state('overwrite')
+    if rename and os.path.exists(old_path):
+        os.remove(old_path)
 
 
 @app.post("/api/save-as", response_class=PlainTextResponse, status_code=200)
-async def save_as_handler(input: SaveAsInput):
+async def save_as_handler(input: SaveAsOrRenameInput):
     """
     Save the state to a new file
     """
     try:
-
-        # Sync the state file
-        context = default_context()
-        state_loader = context.state_loader
-        if state_loader.cloud_mode:
-            raise RecceException('Cloud mode does not support save-as')
-
-        state_loader.state_file = input.filename
-        context.sync_state('overwrite')
-
+        saveas_or_rename(input, rename=False)
     except RecceException as e:
         raise HTTPException(status_code=400, detail=e.message)
 
 
-class SaveAsInput(BaseModel):
-    filename: Optional[str] = None
-    overwrite: Optional[bool] = False
-
-
 @app.post("/api/rename", response_class=PlainTextResponse, status_code=200)
-async def rename_handler(input: SaveAsInput):
+async def rename_handler(input: SaveAsOrRenameInput):
     """
     Rename the state to a new file
     """
-    import os
-
     try:
-        context = default_context()
-        state_loader = context.state_loader
-        if state_loader.cloud_mode:
-            raise RecceException('Cloud mode does not support rename')
-        if state_loader.state_file is None:
-            raise RecceException('No state file to rename')
-
-        old_path = state_loader.state_file
-        new_filename = input.filename
-        old_dir = os.path.dirname(state_loader.state_file)
-        old_filename = os.path.basename(state_loader.state_file)
-        new_path = os.path.join(old_dir, new_filename)
-
-        if os.path.dirname(new_filename):
-            raise RecceException('The new filename should not contain directory')
-        if not new_filename.endswith('.json'):
-            raise RecceException('The new filename should end with .json')
-        if old_filename == input.filename:
-            raise RecceException('The new filename is the same as the current filename')
-        if os.path.exists(new_path) and not input.overwrite:
-            raise HTTPException(status_code=409, detail=f'The file {new_filename} already exists')
-
-        state_loader.state_file = new_path
-        context.sync_state('overwrite')
-
-        if os.path.exists(old_path):
-            os.remove(old_path)
+        saveas_or_rename(input, rename=True)
     except RecceException as e:
         raise HTTPException(status_code=400, detail=e.message)
